# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface LoginResponse {
  token: string;
  username: string;
}

export interface MeResponse {
  username: string;
}

export interface SiteSetting {
  key: string;
  /** @nullable */
  value: string | null;
  type: string;
  label: string;
  section: string;
  /** @nullable */
  hint?: string | null;
  updatedAt: string;
}

export interface PublicSettingsResponse {
  [key: string]: { [key: string]: string | null };
}

export interface UpdateSettingBody {
  key: string;
  value: string;
}

export interface CreateLeadBody {
  name: string;
  email: string;
  phone?: string;
  capitalRange?: string;
  investmentStage?: string;
}

export interface Lead {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  capitalRange?: string | null;
  /** @nullable */
  investmentStage?: string | null;
  source: string;
  createdAt: string;
}

export interface AdminStats {
  leadsToday: number;
  leadsTotal: number;
  imagesCount: number;
  /** @nullable */
  lastUpdate?: string | null;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type UploadSettingImageBody = {
  key: string;
  image: Blob;
};

export type GetAdminLeadsParams = {
  capitalRange?: string;
  dateFrom?: string;
  dateTo?: string;
};

export type ExportLeadsCsvParams = {
  capitalRange?: string;
  dateFrom?: string;
  dateTo?: string;
};