# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: settings
    description: Site settings
  - name: leads
    description: Lead management
  - name: admin
    description: Admin operations
  - name: storage
    description: Object storage

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Not authenticated

  /settings:
    get:
      operationId: getPublicSettings
      tags: [settings]
      summary: Get all public site settings grouped by section
      responses:
        "200":
          description: Settings grouped by section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSettingsResponse"

  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get all settings with metadata
      responses:
        "200":
          description: Settings with admin metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SiteSetting"
    put:
      operationId: updateSetting
      tags: [admin]
      summary: Update a text setting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingBody"
      responses:
        "200":
          description: Setting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSetting"
        "404":
          description: Setting not found

  /admin/settings/image:
    post:
      operationId: uploadSettingImage
      tags: [admin]
      summary: Upload an image for a setting
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                key:
                  type: string
                image:
                  type: string
                  format: binary
              required:
                - key
                - image
      responses:
        "200":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSetting"

  /leads:
    post:
      operationId: createLead
      tags: [leads]
      summary: Submit a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"

  /admin/leads:
    get:
      operationId: getAdminLeads
      tags: [admin]
      summary: List all leads with optional filters
      parameters:
        - name: capitalRange
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

  /admin/leads/export:
    get:
      operationId: exportLeadsCsv
      tags: [admin]
      summary: Export leads as CSV
      parameters:
        - name: capitalRange
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard statistics
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
        username:
          type: string
      required:
        - token
        - username

    MeResponse:
      type: object
      properties:
        username:
          type: string
      required:
        - username

    SiteSetting:
      type: object
      properties:
        key:
          type: string
        value:
          type: ["string", "null"]
        type:
          type: string
        label:
          type: string
        section:
          type: string
        hint:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - key
        - value
        - type
        - label
        - section
        - updatedAt

    PublicSettingsResponse:
      type: object
      additionalProperties:
        type: object
        additionalProperties:
          type: ["string", "null"]

    UpdateSettingBody:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value

    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        capitalRange:
          type: string
        investmentStage:
          type: string
      required:
        - name
        - email

    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        capitalRange:
          type: ["string", "null"]
        investmentStage:
          type: ["string", "null"]
        source:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - source
        - createdAt

    AdminStats:
      type: object
      properties:
        leadsToday:
          type: integer
        leadsTotal:
          type: integer
        imagesCount:
          type: integer
        lastUpdate:
          type: ["string", "null"]
          format: date-time
      required:
        - leadsToday
        - leadsTotal
        - imagesCount

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath