# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  username: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  username: zod.string(),
});

/**
 * @summary Get all public site settings grouped by section
 */
export const GetPublicSettingsResponse = zod.record(
  zod.string(),
  zod.record(zod.string(), zod.string().nullable()),
);

/**
 * @summary Get all settings with metadata
 */
export const GetAdminSettingsResponseItem = zod.object({
  key: zod.string(),
  value: zod.string().nullable(),
  type: zod.string(),
  label: zod.string(),
  section: zod.string(),
  hint: zod.string().nullish(),
  updatedAt: zod.coerce.date(),
});
export const GetAdminSettingsResponse = zod.array(GetAdminSettingsResponseItem);

/**
 * @summary Update a text setting
 */
export const UpdateSettingBody = zod.object({
  key: zod.string(),
  value: zod.string(),
});

export const UpdateSettingResponse = zod.object({
  key: zod.string(),
  value: zod.string().nullable(),
  type: zod.string(),
  label: zod.string(),
  section: zod.string(),
  hint: zod.string().nullish(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Upload an image for a setting
 */
export const UploadSettingImageBody = zod.object({
  key: zod.string(),
  image: zod.any(),
});

export const UploadSettingImageResponse = zod.object({
  key: zod.string(),
  value: zod.string().nullable(),
  type: zod.string(),
  label: zod.string(),
  section: zod.string(),
  hint: zod.string().nullish(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Submit a new lead
 */
export const CreateLeadBody = zod.object({
  name: zod.string(),
  email: zod.string().email(),
  phone: zod.string().optional(),
  capitalRange: zod.string().optional(),
  investmentStage: zod.string().optional(),
});

/**
 * @summary List all leads with optional filters
 */
export const GetAdminLeadsQueryParams = zod.object({
  capitalRange: zod.coerce.string().optional(),
  dateFrom: zod.coerce.string().optional(),
  dateTo: zod.coerce.string().optional(),
});

export const GetAdminLeadsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  phone: zod.string().nullish(),
  capitalRange: zod.string().nullish(),
  investmentStage: zod.string().nullish(),
  source: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetAdminLeadsResponse = zod.array(GetAdminLeadsResponseItem);

/**
 * @summary Export leads as CSV
 */
export const ExportLeadsCsvQueryParams = zod.object({
  capitalRange: zod.coerce.string().optional(),
  dateFrom: zod.coerce.string().optional(),
  dateTo: zod.coerce.string().optional(),
});

/**
 * @summary Dashboard statistics
 */
export const GetAdminStatsResponse = zod.object({
  leadsToday: zod.number(),
  leadsTotal: zod.number(),
  imagesCount: zod.number(),
  lastUpdate: zod.coerce.date().nullish(),
});

/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});