# lib/api-zod/src/generated/types/createLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLeadBody {
  name: string;
  email: string;
  phone?: string;
  capitalRange?: string;
  investmentStage?: string;
}