# lib/api-zod/src/generated/types/siteSetting.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSetting {
  key: string;
  /** @nullable */
  value: string | null;
  type: string;
  label: string;
  section: string;
  /** @nullable */
  hint?: string | null;
  updatedAt: Date;
}